Update PwaCollector and exported_files workflow
The PwaCollector class is updated to import Data from Symfony's VarDumper and modify the return type of the getData method. The isInstallable method is also refactored for a cleaner return statement. The exported_files workflow in the .github folder is altered to additionally exclude the 'templates' directory from the generated tar file.