Features/cross origin credentials (#172)
* Update bootstrap path and add manifest credential use

The bootstrap path in phpunit.xml.dist has been updated to 'tests/autoload.php'. A useCredentials option has been added to the PWA manifest allowing the manifest to be fetched with credentials, enhancing website security. Also, an autoload.php file has been added to facilitate testing.
In the two updated classes (ManifestCompileEventListener and PwaDevServerSubscriber), the constant 'useCredentials' has been added to the IGNORED_ATTRIBUTES option of the Symfony AbstractNormalizer, which prevents serialization of this specific attribute. This adjustment was made to improve the efficiency of our serialization process.