Add nonce support to PwaRuntime
The update includes the use of Nelmio SecurityBundle for Content Security Policy management. Specifically, it checks if the 'nonce' attribute exists in the PwaRuntime and if not, it retrieves a nonce from ContentSecurityPolicyListener for 'script'. This ensures the secure execution of scripts.
In addition, if the 'nonce' attribute is set with the value 'false', the attribute is disabled.